+++ packages/bridge/lib/pi/integrations/phoenix_iconify.ex
@@ -1,9 +1,0 @@
-defmodule Pi.Integrations.PhoenixIconify do
-  @moduledoc "PhoenixIconify-specific project status discovery."
-
-  @behaviour Pi.Integration
-
-  def name, do: :phoenix_iconify
-
-  def statuses, do: []
-end