+++ packages/bridge/lib/pi/integrations/phoenix_replay.ex
@@ -1,9 +1,0 @@
-defmodule Pi.Integrations.PhoenixReplay do
-  @moduledoc "PhoenixReplay-specific project status discovery."
-
-  @behaviour Pi.Integration
-
-  def name, do: :phoenix_replay
-
-  def statuses, do: []
-end