+++ packages/bridge/lib/pi/integrations/volt.ex
@@ -1,9 +1,0 @@
-defmodule Pi.Integrations.Volt do
-  @moduledoc "Volt-specific project status discovery."
-
-  @behaviour Pi.Integration
-
-  def name, do: :volt
-
-  def statuses, do: []
-end